requirements.txt
tensorflow-cpu==2.8.0
numpy==1.20
protobuf==3.17.3
joblib==1.0.1
APScheduler==3.9.1
SQLAlchemy==1.4.22
PyMySQL==1.0.2
requests==2.26.0
cryptography==3.4.7   